Update puppet-agent patches for 6.12.0
This commit rebases the puppet-agent patchset for version 6.12.0.
The main change that prompted this rebase was the addition of Fedora
versions to case statements that skipped pl-built-tools. This caused
conflicts in patch #3. The patch has been re-worked to add Debian armhf
as it's own isolated case rather than trying to consolidate cases.