Add script for fetching Vanagon component versions
This commit adds scripts/component-version.sh which uses curl and jq to
fetch the version of Vanagon components that go into a particular
project build. Only works for components where the version number is
stored in a .json file.

The build workflow is also re-factored to use the script and centralize
some environment variables. This is done by using "logging commands" to
set values in the GitHub action environment:

  https://help.github.com/en/actions/reference/development-tools-for-github-actions#logging-commands