Work around --flat being broken for download-ver
The `--flat` flag for `jfrog bt download-ver` is not having its
advertised effect as the command is still re-creating the directory
structure from BinTray instead of just dumping the files in the
designated output directory.

This commit downloads the files to a staging location and then
moves them into place.