Add ability to rebuild versions
This commit adds a `rebuild` parameter that can be set in the dispatch
event data used to trigger the build. When set, the job will build
puppet-runtime and puppet-agent from scratch and re-upload the results
to BinTray.

This is useful when a previous build created an artifact that failed
tests and needs to be modified. Attempting to overwrite published
artifacts will fail.