Skip weighted_cached_custom_facts test for facter
This commit adds weighted_cached_custom_facts.rb to the list of tests
that are skipped when testing facter. This test consistently produces
an empty cache of data, `{}`, instead of caching as it is supposed to.
This behavior is re-producible in released x86_64 packages for
Debian 10, which points to a bug in Facter.