Skip facter/facter-ng comparison test
The `puppet-agent` acceptance test that compares Facter 4 (facter-ng)
output to Facter 3 still fails for Raspbian. However, this is not a
fault in the build. Rather, Facter 4 prefers `/etc/os-release` as a
source of distro info and thus properly reports "Raspbian" as the
distro name. Facter 3 reports "Debian" as it assigns a higher
precedence to the `/etc/debian_version` file.