Remove JFrog Bintra integration
This commit updates the puppet-aent build pipeline to remove
logic related to caching and shipping binaries to JFrog Bintray
as the service shut down on May 1st, 2020.

The output of the GitHub build pipeline is now stored locally
as an artifact. Artifacts will eventually expire and be deleted,
but the output of a successful build can be manually transferred
to a permanent location.

Automatic shipping of packages to a Debian repository may return
in the future when a new hosting solution is determined.