Add Bolt task to build puppet-agent
This commit adds a `build_component` tasks to the Bolt project
in `resources/bolt` that can build the `puppet-runtime` and
`puppet-agent` packages on an ARM host. The Raspberry Pi OS
workflow is updated to use this new task.

This commit also adds patches required for native builds of
puppet-agent 6.22.0 and newer.