Set platform triple in puppet-runtime native build
This commit adds one patch to puppet-runtime for native
builds that passes the platform triple as arguments to
`--host` and `--target` for Autotools `./configure` scripts.

This is important for building Ruby on something that is
not RPi 4 hardware, as the resulting binary can end up
with `linux-armv8l` if `--host` is not set and can
end up dropping `gnu` from `arm-linux-gnueabihf` to
end up with `arm-linux-eabihf` if `--target` is not set.